fix: stabilize threat indicator and apply selected protection mode
PROBLEMS FIXED:
- Indicator stayed green with 3 faces or turned red with 1 face
- PARANOIA mode selected in settings was never applied (always DISCRETE)
- Oscillations between colors (greenyellowred)

CHANGES:

1. Camera Scoring (ThreatScorer.kt):
   - 1 face: 5% (was 10%) - user alone
   - 2 faces: 30% (was 35%) - real threat
   - 3 faces: 45% (was 40%) - high threat
   - 4+ faces: 50% max - critical threat
   - More logical looking-at-screen detection (35% if multiple faces)

2. Audio Scoring (ThreatScorer.kt):
   - Adjusted thresholds: < 40dB = safe, 65-75dB = monitoring, > 85dB = threat
   - Speech detection bonus: 30% (was 40%)

3. Indicator Logic (PrivacyGuardService.kt):
   - REMOVED hybrid logic (was forcing red inconsistently)
   - NOW: Simple mapping ThreatLevel  IndicatorState
   - CRITICAL/HIGH  Red, MEDIUM  Yellow, LOW/NONE  Green

4. ThreatLevel Thresholds (SensorDataFusion.kt):
   - Widened zones for stability:
   - NONE: 0-24, LOW: 25-44, MEDIUM: 45-64, HIGH: 65-84, CRITICAL: 85+

5. Stability Filter (ThreatAssessmentEngine.kt):
   - Added .scan() to count stable readings
   - Requires 2 consecutive samples (200ms) before emitting
   - CRITICAL level bypasses filter (immediate response)
   - Debounce increased: 50ms  100ms
   - distinctUntilChanged granularity: score/10  score/15

6. Protection Mode Application (PrivacyGuardService.kt):
   - BUG FIX: Service now reads saved mode from SharedPreferences
   - Was hardcoded to DISCRETE, now respects user selection
   - PARANOIA mode now actually applied (threshold 20 vs 75)

7. PARANOIA Weights (ThreatModels.kt):
   - Camera: 35%  50% (face detection priority)
   - Audio: 30%  35%
   - Motion: 25%  10% (reduced noise)
   - Proximity: 10%  5%

RESULTS:
- 3 faces + music: 52%  69% (MEDIUMHIGH, YellowRed)
- 1 face alone: stays green
- 2 faces + conversation: yellow
- No more flickering between colors
- PARANOIA mode 3.75x more sensitive than DISCRETE

Docs: FIX_INDICATOR_STABILITY.md, BUG_MODE_PARANOIA.md

+++ app/src/test/java/com/privacyguard/assessment/ThreatScorerTest.kt
@@ -254,7 +254,7 @@ class ThreatScorerTest {
             contributions.cameraScore * contributions.cameraWeight +
             contributions.audioScore * contributions.audioWeight
         ) * 100
-        assertEquals(expectedScore.toInt(), finalScore, 5) // Tolérance de 5
+        assertEquals(expectedScore.toInt().toLong(), finalScore.toLong(), 5L) // Tolérance de 5
     }
     
     // ==================== Tests de confiance ====================
